+++ wss-server-core/src/main/java/de/uniba/dsg/wss/auth/EmployeeUserDetailsService.java
@@ -1,6 +1,7 @@
 package de.uniba.dsg.wss.auth;
 
-import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -12,14 +13,35 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
  */
 public abstract class EmployeeUserDetailsService implements UserDetailsService {
 
-  public EmployeeUserDetailsService() {}
+  private final AuthorityMapping authorityMapping;
+
+  public EmployeeUserDetailsService(AuthorityMapping authorityMapping) {
+    this.authorityMapping = authorityMapping;
+  }
 
   @Override
   public abstract EmployeeUserDetails loadUserByUsername(String username)
       throws UsernameNotFoundException;
 
-  protected EmployeeUserDetails createWithDefaultRole(String username, String password) {
-    return new EmployeeUserDetails(
-        username, password, List.of(new SimpleGrantedAuthority(Role.TERMINAL_USER.prefixedName())));
+  /**
+   * Creates a new {@code EmployeeUserDetails} instance with the given parameters. Note that this
+   * method uses the given {@code role} to determine the effective authorities of the user.
+   *
+   * <p>Implementations of this type of service may use this method in their implementation of
+   * {@link #loadUserByUsername(String)} to create the user details object to be returned.
+   *
+   * @param username the unique username identifying an employee account
+   * @param password the corresponding password
+   * @param role one of the roles defined in {@link Roles}, can be prefixed
+   * @return the newly created user details object
+   * @throws IllegalArgumentException if the given role is not one of the roles defined in {@link
+   *     Roles}
+   */
+  protected EmployeeUserDetails createUserDetails(String username, String password, String role) {
+    Set<SimpleGrantedAuthority> effectiveAuthorities =
+        authorityMapping.getEffectiveAuthorities(role).stream()
+            .map(SimpleGrantedAuthority::new)
+            .collect(Collectors.toSet());
+    return new EmployeeUserDetails(username, password, effectiveAuthorities);
   }
 }